added models ui

+++ controllers/modelsController.js
@@ -2,8 +2,42 @@ const Models = require('../models/modelsModel');
 
 const getModels = async (req, res, next) => {
   try {
-    const response = await Models.retrieveAll();
-    res.send(response);
+    const response = (await Models.retrieveAll()).sort((a, b) => new Date(b.id) - new Date(a.id));
+    // console.log(response);
+    res.render('index', {
+      windowTitle: 'Models | RK Inventory',
+      documentTitle: 'Royal Kludge Inventory',
+      header: {
+        icons: {
+          hamburger: '<svg xmlns="http://www.w3.org/2000/svg" height="30px" viewBox="0 -960 960 960" width="30px" fill="#f5f5f5"> <path d="M120-680v-80h720v80H120Zm0 480v-80h720v80H120Zm0-240v-80h720v80H120Z"/> </svg>',
+          close: '<svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 -960 960 960" width="24px" fill="#f5f5f5"><path d="m256-200-56-56 224-224-224-224 56-56 224 224 224-224 56 56-224 224 224 224-56 56-224-224-224 224Z"/></svg>',
+        }
+      },
+      content: {
+        name: 'model',
+        data: response
+      }
+    });
+  } catch (error) {
+    next(error);
+  }
+}
+
+const renderCreatePage = async (req, res, next) => {
+  try {
+    res.render('index', {
+      windowTitle: 'New Model | RK Inventory',
+      documentTitle: 'Royal Kludge Inventory',
+      header: {
+        icons: {
+          hamburger: '<svg xmlns="http://www.w3.org/2000/svg" height="30px" viewBox="0 -960 960 960" width="30px" fill="#f5f5f5"> <path d="M120-680v-80h720v80H120Zm0 480v-80h720v80H120Zm0-240v-80h720v80H120Z"/> </svg>',
+          close: '<svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 -960 960 960" width="24px" fill="#f5f5f5"><path d="m256-200-56-56 224-224-224-224 56-56 224 224 224-224 56 56-224 224 224 224-56 56-224-224-224 224Z"/></svg>',
+        }
+      },
+      content: {
+        name: 'newModel'
+      }
+    });
   } catch (error) {
     next(error);
   }
@@ -11,9 +45,27 @@ const getModels = async (req, res, next) => {
 
 const getModelById = async (req, res, next) => {
   try {
+    const { auth } = req.query;
+    if (auth !== process.env.SECRET_PASSWORD) {
+      return res.status(401).send('Unauthorized: Invalid password');
+    }
     const { id } = req.params;
-    const response = await Models.retrieveModelById(id);
-    res.send(response);
+    const { model } = await Models.retrieveModelById(id);
+    // console.log(model);
+    res.render('index', {
+      windowTitle: 'Edit Model | RK Inventory',
+      documentTitle: 'Royal Kludge Inventory',
+      header: {
+        icons: {
+          hamburger: '<svg xmlns="http://www.w3.org/2000/svg" height="30px" viewBox="0 -960 960 960" width="30px" fill="#f5f5f5"> <path d="M120-680v-80h720v80H120Zm0 480v-80h720v80H120Zm0-240v-80h720v80H120Z"/> </svg>',
+          close: '<svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 -960 960 960" width="24px" fill="#f5f5f5"><path d="m256-200-56-56 224-224-224-224 56-56 224 224 224-224 56 56-224 224 224 224-56 56-224-224-224 224Z"/></svg>',
+        }
+      },
+      content: {
+        name: 'editModel',
+        model
+      }
+    });
   } catch (error) {
     next(error);
   }
@@ -21,9 +73,9 @@ const getModelById = async (req, res, next) => {
 
 const addModel = async (req, res, next) => {
   try {
-    const newProduct = { name, layout, connectivity, switch_support } = req.body;
-    const response = await Models.createModel(newProduct);
-    res.send(response);
+    const newModel = { name, layout, connectivity, switch_support } = req.body;
+    await Models.createModel(newModel);
+    res.redirect('/models');
   } catch (error) {
     next(error);
   }
@@ -33,8 +85,8 @@ const updateModel = async (req, res, next) => {
   try {
     const { id } = req.params;
     const updatingModel = req.body;
-    const response = await Models.updateModelById(id, updatingModel);
-    res.send(response);
+    await Models.updateModelById(id, updatingModel);
+    res.redirect('/models');
   } catch (error) {
     next(error);
   }
@@ -42,9 +94,13 @@ const updateModel = async (req, res, next) => {
 
 const deleteModel = async (req, res, next) => {
   try {
+    const { auth } = req.body;
+    if (auth !== process.env.SECRET_PASSWORD) {
+      return res.status(401).send('Unauthorized: Invalid password');
+    }
     const { id } = req.params;
-    const response = await Models.deleteModelById(id);
-    res.send(response);
+    await Models.deleteModelById(id);
+    res.redirect('/models');
   } catch (error) {
     next(error);
   }
@@ -55,5 +111,6 @@ module.exports = {
   getModelById,
   addModel,
   updateModel,
-  deleteModel
+  deleteModel,
+  renderCreatePage
 }
